% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_km.R
\name{fit_km}
\alias{fit_km}
\title{Fit Kaplan-Meier survival curves}
\usage{
fit_km(data, time, status, strata = NULL, weights = NULL, conf.type = "log")
}
\arguments{
\item{data}{Data frame containing the variables.}

\item{time}{Name of the time variable.}

\item{status}{Name of the event variable (1 = event, 0 = censor).}

\item{strata}{Optional name of a grouping (strata) variable.}

\item{weights}{Optional case weights.}

\item{conf.type}{Type of confidence interval ("log", "plain", etc).}
}
\value{
A survfit object.
}
\description{
Fit Kaplan-Meier survival curves
}
