% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration_surv.R
\name{plot_calibration_surv}
\alias{plot_calibration_surv}
\title{Plot observed vs predicted survival (calibration curve)}
\usage{
plot_calibration_surv(cal_df)
}
\arguments{
\item{cal_df}{Data frame returned by \code{calibrate_surv()}.}
}
\value{
A \code{ggplot} object showing observed vs predicted survival per group.
}
\description{
Plot observed vs predicted survival (calibration curve)
}
\examples{
\dontrun{
  data(lung, package = "survival")
  lung$status01 <- as.integer(lung$status == 2)
  cfit <- survival::coxph(survival::Surv(time, status01) ~ age + sex, data = lung)
  cal <- calibrate_surv(cfit, lung, "time", "status01", time_point = 365)
  p <- plot_calibration_surv(cal); print(p)
}
}
