% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_weighted.R
\name{survfit_weighted}
\alias{survfit_weighted}
\title{Weighted Kaplan–Meier (with optional strata)}
\usage{
survfit_weighted(
  data,
  time,
  status,
  strata = NULL,
  weights = NULL,
  conf.type = "log"
)
}
\arguments{
\item{data}{data.frame}

\item{time}{Name of time column}

\item{status}{Name of status column (1 = event, 0 = censor)}

\item{strata}{Optional name of a grouping column}

\item{weights}{Optional vector or column-name of case weights}

\item{conf.type}{CI type passed to \code{survival::survfit()}}
}
\value{
A \code{survfit} object
}
\description{
Weighted Kaplan–Meier (with optional strata)
}
